+++ src/main/java/artifacts/integrations/ModMenuConfig.java
@@ -1,7 +1,8 @@
-package artifacts.config;
+package artifacts.integrations;
 
-import io.github.prospector.modmenu.api.ConfigScreenFactory;
-import io.github.prospector.modmenu.api.ModMenuApi;
+import artifacts.config.ModConfig;
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
 import me.shedaniel.autoconfig.AutoConfig;
 
 public class ModMenuConfig implements ModMenuApi {